requirements.txt
psutil==5.9.7
speedtest-cli==2.1.4b1
